psoc-edge: Implemented X-macro from make-pins.
