+++ ports/psoc-edge/Makefile
@@ -585,7 +585,7 @@ MOD_SRC_C += \
 	machine_rtc.c \
 	machine_ipc.c \
 	machine_scb.c \
-	machine_tcpwm.c \
+	tcpwm.c \
 	machine_timer.c \
 	modpsocedge.c \
